+++ src/ad9834_ad9838.rs
@@ -1,4 +1,7 @@
-use {marker, Ad983x, BitFlags, ControlSource, Error, OutputWaveform, SignBitOutput, SpiInterface};
+use {
+    marker, Ad983x, BitFlags, ControlSource, Error, OutputWaveform, SignBitOutput, SpiInterface,
+    SpiWrite,
+};
 
 impl<SPI, CS> Ad983x<SpiInterface<SPI, CS>, marker::Ad9834Ad9838> {
     /// Create a new instance of an AD9834 device.
@@ -14,17 +17,16 @@ impl<SPI, CS> Ad983x<SpiInterface<SPI, CS>, marker::Ad9834Ad9838> {
     }
 }
 
-impl<SPI, CS, E> Ad983x<SpiInterface<SPI, CS>, marker::Ad9834Ad9838>
+impl<CommE, PinE, DI> Ad983x<DI, marker::Ad9834Ad9838>
 where
-    SPI: hal::blocking::spi::Write<u8, Error = E>,
-    CS: hal::digital::OutputPin,
+    DI: SpiWrite<Error = Error<CommE, PinE>>,
 {
     /// Set the output waveform
     ///
     /// Will return `Error::InvalidArgument` for `SquareMsbOfDac` and `SquareMsbOfDacDiv2`
     /// as this is not available on AD9834/AD9838 devices. To set the digital output,
     /// please use
-    pub fn set_output_waveform(&mut self, waveform: OutputWaveform) -> Result<(), Error<E>> {
+    pub fn set_output_waveform(&mut self, waveform: OutputWaveform) -> Result<(), DI::Error> {
         let control;
         match waveform {
             OutputWaveform::Sinusoidal => {
@@ -46,7 +48,7 @@ where
     }
 
     /// Set the digital output
-    pub fn set_sign_bit_output(&mut self, configuration: SignBitOutput) -> Result<(), Error<E>> {
+    pub fn set_sign_bit_output(&mut self, configuration: SignBitOutput) -> Result<(), DI::Error> {
         let control = match configuration {
             SignBitOutput::Disabled => self.control.with_low(BitFlags::OPBITEN),
             SignBitOutput::Comparator => self
@@ -74,7 +76,7 @@ where
     /// Set the control source used for the functions:
     /// frequency register selection, phase register selection,
     /// reset of internal registers, and DAC power-down.
-    pub fn set_control_source(&mut self, source: ControlSource) -> Result<(), Error<E>> {
+    pub fn set_control_source(&mut self, source: ControlSource) -> Result<(), DI::Error> {
         let control = match source {
             ControlSource::Software => self.control.with_low(BitFlags::PIN_SW),
             ControlSource::HardwarePins => self.control.with_high(BitFlags::PIN_SW),