+++ tests/base/mod.rs
@@ -24,9 +24,14 @@ impl BitFlags {
 
 pub struct DummyOutputPin;
 
-impl embedded_hal::digital::OutputPin for DummyOutputPin {
-    fn set_low(&mut self) {}
-    fn set_high(&mut self) {}
+impl embedded_hal::digital::v2::OutputPin for DummyOutputPin {
+    type Error = ();
+    fn set_low(&mut self) -> Result<(), ()> {
+        Ok(())
+    }
+    fn set_high(&mut self) -> Result<(), ()> {
+        Ok(())
+    }
 }
 
 pub fn new_ad9833(